app_nckh9/kaira_chatbot/requirements.txt
streamlit>=1.30.0
nltk>=3.8.1
faiss-cpu>=1.7.4
torch>=2.1.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
transformers>=4.36.0
python-docx>=1.0.1
tenacity>=8.2.3
numpy>=1.24.3
PyQt5>=5.15.0